+++ WSDOT/Models/Shared/Resource.cs
